requirements.txt
tornado
requests
beautifulsoup4
nltk
pycrypto
SQLAlchemy
